add and integrate with dynamodb storage
Changes:
-   implemented dynamodb layer as a storage layer
-   defined the recipes table with recipe_id as the partition_key
-   query by other fields other than the recipe_id will be added later
-   added dynamodb table to cdk
-   increased lambda timeout to allow higher cold start times
-   added support to run a local instance of dynamodb server (port 8000)
-   created a layer to manage app initialization (RecipeApp)
    -   if it's running in sam local, it uses the local version of dynamodb
    -   otherwise, use the temporary memory storage (which will become dynamodb storage once deployed
        to production)
-   documented a limitation in FastAPI that doesn't prevent users from sending unknown [model] fields
-   replaced recipe_id type from uuid to string. An alternative is to use a custom json encoder,
    which I can do in a later stage.
-   make datetime_to_milliseconds return int instead of float (dynamodb doesn't support float)
-   I'm relying on integration tests to test these changes. Given this is an interview test, I chose
    to not implement unit tests for this commit, otherwise I would spend a good amount of time on tests.
    In an usual working environment, I would.

+++ gota-core/src/gota/core/time_converter.py
@@ -2,4 +2,4 @@ import arrow
 
 
 def datetime_to_milliseconds(datetime: arrow.Arrow):
-    return datetime.timestamp() * 1000
+    return int(datetime.timestamp() * 1000)
